Support registraton domain and proper sync
On 10.6 the periodic sync behavior no longer uses synchronize and the
values were being written to the hosting app's persistent domain. Now
all modifications are intercepted and go through our own version of
synchronize.

Also added support for registered defaults in the volatile domain and
posting of a notification when defaults are changed.